# docs/swagger.yaml
definitions:
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Item:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_req.Service:
    properties:
      labels:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      name:
        type: string
      namespace:
        type: string
      ports:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.ServicePort'
        type: array
      selector:
        items:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Item'
        type: array
      type:
        $ref: '#/definitions/v1.ServiceType'
    type: object
  github_com_crazyfrankie_kube-ctl_internal_model_req.ServicePort:
    properties:
      name:
        type: string
      nodePort:
        type: integer
      port:
        type: integer
      targetPort:
        type: integer
    type: object
...
  github_com_crazyfrankie_kube-ctl_internal_model_resp.Service:
    properties:
      age:
        type: integer
      clusterIP:
        type: string
      externalIP:
        items:
          type: string
        type: array
      name:
        type: string
      namespace:
        type: string
      type:
        $ref: '#/definitions/v1.ServiceType'
    type: object
...
  github_com_crazyfrankie_kube-ctl_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  v1.ServiceType:
    enum:
    - ClusterIP
    - NodePort
    - LoadBalancer
    - ExternalName
    type: string
    x-enum-varnames:
    - ServiceTypeClusterIP
    - ServiceTypeNodePort
    - ServiceTypeLoadBalancer
    - ServiceTypeExternalName
...
info:
  contact: {}
paths:
...
  /api/service:
    delete:
      consumes:
      - application/json
      description: 删除一个 Service
      parameters:
      - description: Service 名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除 Service 成功
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 删除 Service
      tags:
      - Service 管理
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下指定Service的详细信息
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: Service 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回Service的详细信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Service'
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取Service详情
      tags:
      - Service 管理
    post:
      consumes:
      - application/json
      description: 创建新的 Service 或更新已存在的 Service
      parameters:
      - description: Service 配置信息
        in: body
        name: pod
        required: true
        schema:
          $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_req.Service'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功，返回成功消息
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "400":
          description: 参数错误(code=20001)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 创建或更新 Service
      tags:
      - Service 管理
  /api/service/list:
    get:
      consumes:
      - application/json
      description: 获取指定命名空间下指定Service的列表
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 关键词
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回Service的列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_internal_model_resp.Service'
                  type: array
              type: object
        "500":
          description: 系统错误(code=30000)
          schema:
            $ref: '#/definitions/github_com_crazyfrankie_kube-ctl_pkg_response.Response'
      summary: 获取Service列表
      tags:
      - Service 管理
...
swagger: "2.0"